requirements.txt
torch
torchvision
numpy
opencv-python-headless
fastapi
uvicorn
pillow